edits to comment
