+++ src/components/burger-ingredients/BurgerIngredients.jsx
@@ -4,6 +4,7 @@ import {CurrencyIcon} from '@ya.praktikum/react-developer-burger-ui-components';
 import {Counter} from '@ya.praktikum/react-developer-burger-ui-components';
 import ingredientsStyles from './burger-ingredients.module.css';
 import PropTypes from "prop-types";
+import types from "../../utils/types";
 import BurgerConstructor from "../burger-constructor/BurgerConstructor";
 
 class BurgerIngredients extends Component {
@@ -87,21 +88,8 @@ class BurgerIngredients extends Component {
     }
 }
 
-const burgerIngredientsPropTypes = PropTypes.shape({
-    _id: PropTypes.string.isRequired,
-    name: PropTypes.string.isRequired,
-    type: PropTypes.string.isRequired,
-    proteins: PropTypes.number,
-    fat: PropTypes.number,
-    carbohydrates: PropTypes.number,
-    calories: PropTypes.number,
-    price: PropTypes.number.isRequired,
-    image: PropTypes.string,
-    image_mobile: PropTypes.string,
-    image_large: PropTypes.string,
-    __v: PropTypes.number,
-});
-
-BurgerIngredients.propTypes = burgerIngredientsPropTypes.isRequired;
+BurgerIngredients.propTypes = {
+    data: PropTypes.arrayOf(types.isRequired)
+}
 
 export default BurgerIngredients;
